Add more exotic trombones, saxhorns, and tubas
"Trombone in Bb", "tuba in Eb", and "tuba in Bb" are transposing
counterparts to their respective standard types following the
convention of British brass band notation.

"Alto trombone" and "bass trombone" are higher- and lower-pitched
variants to complement the standard (tenor) trombone type.

"Alto horn" and "tenor horn" are alternate names for the same
instrument in American and British English, respectively. They are
included to complement the existing "baritone" (horn) type.